# packages/api-contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Verdery API
  version: 1.0.0
  summary: Garden mapping, care records, and recommendations
  description: |
    The external contract shared by the native Apple and web applications.

    This document is the machine-readable source of truth for paths, schemas,
    error envelopes, pagination, idempotency, and concurrency headers. Generated
    clients are wrapped by handwritten application gateways and are never edited
    by hand.

    Conventions:

    - Paths use plural lowercase nouns; JSON fields use `camelCase`.
    - Identifiers are opaque UUIDv7 strings.
    - Timestamps are RFC 3339 in UTC.
    - Measurements are SI with explicit unit metadata where ambiguity exists.
    - Geometry always travels inside an envelope naming its coordinate space,
      because GeoJSON alone cannot say whether coordinates are local metres or
      longitude and latitude.

    Source: architecture/api-design.md.
  license:
    name: UNLICENSED
    identifier: LicenseRef-Proprietary
servers:
  - url: /v1
    description: Version 1
...
paths:
...
  /gardens/{gardenId}/today:
    parameters:
      - name: gardenId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Uuid'

    get:
      operationId: getTodayRecommendations
      tags: [Recommendations]
      summary: Get the Today recommendation set
      description: |
        The garden's small prioritized set of actionable recommendations:
        candidates in `eligible`/`presented` whose validity window covers
        now, ordered by priority score descending — the score is re-derived
        server-side from each candidate's STORED priority factors as the
        clamped-to-[0,100] sum of their integer `contribution`s, so the
        stored factors alone reproduce and explain the rank — with sooner
        `windowEnd` first (nulls last) and candidate id as tie-breakers.

        This query records FIRST PRESENTATION as a deliberate side effect:
        an `eligible` candidate included in the response transitions to
        `presented` (stamping `presentedAt` and bumping `revision`) in the
        same transaction that read it, because inclusion in a returned
        Today response IS the presentation fact. The transition happens
        only on first inclusion, so repeating the request is safe — later
        reads find `presented` and write nothing. Candidates beyond the
        `limit` cap are not returned and not marked.

        Source: implementation-plan.md work package P7-BE-01;
        recommendations-and-ai.md section 6.
      parameters:
        - name: limit
          in: query
          required: false
          description: Maximum items to return. Defaults to 10.
          schema:
            type: integer
            minimum: 1
            maximum: 25
            default: 10
      responses:
        '200':
          description: The prioritized Today set.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodayResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /gardens/{gardenId}/recommendations/{recommendationId}/complete:
    parameters:
      - name: gardenId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Uuid'
      - name: recommendationId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Uuid'

    post:
      operationId: completeRecommendation
      tags: [Recommendations]
      summary: Complete a recommendation
      description: |
        Records `completed` feedback and transitions the candidate
        `presented -> completed` in one transaction. Only legal while
        `presented`.

        Source: implementation-plan.md work package P7-BE-01.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
      responses:
        '200':
          description: The completed recommendation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recommendation'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
  /gardens/{gardenId}/recommendations/{recommendationId}/postpone:
    parameters:
      - name: gardenId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Uuid'
      - name: recommendationId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Uuid'

    post:
      operationId: postponeRecommendation
      tags: [Recommendations]
      summary: Postpone a recommendation
      description: |
        Records `postponed` feedback (with the caller's optional
        `postponedUntil` horizon) and transitions the candidate
        `presented -> postponed` in one transaction. `postponed` is
        TERMINAL: re-surfacing is the rule engine's job on a later
        evaluation — after the horizon (or the rule's recurrence interval
        when none was given) it may generate a NEW candidate referencing
        this one via `supersedesCandidateId`, preserving this record's
        evidence and feedback unmodified. Only legal while `presented`.

        Source: implementation-plan.md work package P7-BE-01;
        recommendations-and-ai.md section 6.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostponeRecommendationRequest'
      responses:
        '200':
          description: The postponed recommendation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recommendation'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
  /gardens/{gardenId}/recommendations/{recommendationId}/dismiss:
    parameters:
      - name: gardenId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Uuid'
      - name: recommendationId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Uuid'

    post:
      operationId: dismissRecommendation
      tags: [Recommendations]
      summary: Dismiss a recommendation
      description: |
        Records `dismissed` feedback and transitions the candidate
        `presented -> rejected` in one transaction — the feedback verb is
        FR-24's, the state is the lifecycle's own vocabulary. Only legal
        while `presented`.

        Source: implementation-plan.md work package P7-BE-01.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
      responses:
        '200':
          description: The dismissed (now `rejected`) recommendation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recommendation'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
  /gardens/{gardenId}/recommendations/{recommendationId}/mark-irrelevant:
    parameters:
      - name: gardenId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Uuid'
      - name: recommendationId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Uuid'

    post:
      operationId: markRecommendationIrrelevant
      tags: [Recommendations]
      summary: Report a recommendation as not relevant
      description: |
        Appends `irrelevant` feedback WITHOUT a lifecycle transition or
        revision bump — the explicit quality signal that accompanies or
        follows a dismissal. Legal on a still-visible `presented`
        candidate or on an already-`rejected` one; the response echoes the
        unchanged recommendation.

        Source: implementation-plan.md work package P7-BE-01;
        technical-specification.md FR-3.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
      responses:
        '200':
          description: The unchanged recommendation the feedback now annotates.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recommendation'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
  /gardens/{gardenId}/recommendations/{recommendationId}/convert-to-task:
    parameters:
      - name: gardenId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Uuid'
      - name: recommendationId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Uuid'

    post:
      operationId: convertRecommendationToTask
      tags: [Recommendations]
      summary: Convert a recommendation into a task
      description: |
        One transaction: the candidate transitions
        `presented -> completed` (converting is acting on it — the work
        now lives on the task list), a `completed` feedback row is
        appended, and a task is created with `source: 'suggested'` and
        `originRecommendationId` set together (the database enforces
        exactly that pairing), `status: 'planned'`, the rule version's own
        action title as the task title, the stored deterministic
        explanation as the task notes, and the candidate's target,
        urgency, and validity window carried over. The converted task's
        `originRecommendationId` is what distinguishes a
        conversion-completion from a did-it-now completion in the outcome
        history. Only legal while `presented`.

        Source: implementation-plan.md work package P7-BE-01;
        technical-specification.md FR-25.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
      responses:
        '201':
          description: The created task and the completed recommendation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConvertRecommendationToTaskResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
...
components:
  securitySchemes:
    firebaseIdToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Firebase ID token presented by the native client.

        Source: architecture/identity-and-authorization.md, section
        "4. Native Authentication Flow".
    sessionCookie:
      type: apiKey
      in: cookie
      name: __session
      description: |
        HTTP-only Firebase session cookie used by the web client. Cookie
        authenticated mutations additionally require CSRF protection.

        Source: architecture/identity-and-authorization.md, section
        "5. Web Session Flow".
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Client-generated UUIDv7. The same key with a semantically identical
        request returns the original result. The same key with a different
        command is rejected with `request.idempotency.key_reused`.
      schema:
        $ref: '#/components/schemas/Uuid'
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: |
        Expected revision of the target resource, quoted. A stale value is
        rejected rather than silently overwriting a newer state.
      schema:
        type: string
        examples: ['"7"']
...
  schemas:
    Uuid:
      type: string
      format: uuid
      description: Opaque UUIDv7 identifier. Time ordering is an index property, not an authoritative creation time.
      pattern: '^[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$'
      examples: ['019827ab-4c1d-7e3f-9a2b-5c6d7e8f9a0b']
    Timestamp:
      type: string
      format: date-time
      description: RFC 3339 timestamp in UTC.
      examples: ['2026-07-22T09:15:00Z']
    Revision:
      type: integer
      format: int64
      minimum: 1
      description: Monotonically increasing server revision within one object's identity.
...
    ErrorDetail:
      type: object
      additionalProperties: false
      required: [code]
      properties:
        code:
          type: string
          description: Stable machine-readable detail code.
        pointer:
          type: string
          description: JSON Pointer to the offending member, where the error is structural.
        parameters:
          type: object
          additionalProperties: true
          description: Values interpolated into the localized message. Never free-form prose.
    Error:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, correlationId, retryable]
          properties:
            code:
              type: string
              description: |
                Stable dotted code. Clients localize known errors by this value
                rather than by displaying `message`.
              examples: ['garden.geometry.stale_revision']
            message:
              type: string
              description: Safe fallback text in English. Not a localization source.
            correlationId:
              type: string
              description: Identifier for correlating this response with server telemetry.
            details:
              type: array
              items:
                $ref: '#/components/schemas/ErrorDetail'
            retryable:
              type: boolean
              description: Whether repeating the identical request may succeed later.
...
    TaskTargetKind:
      type: string
      enum: [garden, garden_area, plant]
...
    TaskUrgency:
      type: string
      enum: [low, normal, high, urgent]
...
    RecommendationState:
      type: string
      enum: [generated, eligible, presented, completed, postponed, rejected, expired, superseded]
      description: |
        Section 6's candidate lifecycle. Responses here only ever carry
        `presented` (Today items are marked presented on first inclusion)
        or the state a command produced (`completed`, `postponed`,
        `rejected`); the full vocabulary is declared so clients can render
        any candidate a future surface exposes.
    RecommendationSafetyTier:
      type: string
      enum: [ordinary_care, elevated_risk]
      description: |
        Section 13's tiers, minus `restricted`: a restricted-tier candidate
        is physically impossible (database CHECK + engine type), so the
        contract does not pretend one could appear.
...
    RecommendationPriorityFactor:
      type: object
      additionalProperties: false
      required: [kind, contribution, basis]
      description: |
        One stored, explainable priority input: the candidate's score is
        the clamped-to-[0,100] sum of these integer contributions — the
        factors alone reproduce the rank. `basis` names the facts the
        contribution derives from (uncertainty signals such as low
        confidence or a stale-weather label live here and in the evidence
        values).
      properties:
        kind:
          $ref: '#/components/schemas/RecommendationPriorityFactorKind'
        contribution:
          type: integer
          minimum: -100
          maximum: 100
        basis:
          type: object
          additionalProperties: true
    RecommendationEvidence:
      type: object
      additionalProperties: false
      required:
        [
          id,
          kind,
          sourceObservationId,
          sourceTaskId,
          sourcePlantId,
          sourceWeatherRecordId,
          factKey,
          factValue,
        ]
      description: |
        One structured fact the recommendation rests on — FR-24's
        "Evidence used". Reference kinds carry exactly their own source id;
        context kinds carry none. `factValue` is the generation-time value
        snapshot, `null` when the referenced row itself is the value.
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        kind:
          $ref: '#/components/schemas/RecommendationEvidenceKind'
        sourceObservationId:
          oneOf:
            - $ref: '#/components/schemas/Uuid'
            - type: 'null'
        sourceTaskId:
          oneOf:
            - $ref: '#/components/schemas/Uuid'
            - type: 'null'
        sourcePlantId:
          oneOf:
            - $ref: '#/components/schemas/Uuid'
            - type: 'null'
        sourceWeatherRecordId:
          oneOf:
            - $ref: '#/components/schemas/Uuid'
            - type: 'null'
        factKey:
          type: string
          minLength: 1
          maxLength: 200
        factValue: {}
    Recommendation:
      type: object
      additionalProperties: false
      required:
        [
          id,
          gardenId,
          ruleKey,
          ruleVersion,
          careCategory,
          safetyTier,
          state,
          urgency,
          targetKind,
          targetGardenAreaMapObjectId,
          targetPlantId,
          windowStart,
          windowEnd,
          explanation,
          supersedesCandidateId,
          presentedAt,
          revision,
          createdAt,
          updatedAt,
        ]
      description: |
        The command-response summary of one recommendation candidate.
        `explanation` is the deterministic reason rendered and stored at
        generation time — never re-rendered against current facts.
        `revision` feeds the next command's `If-Match`.
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        gardenId:
          $ref: '#/components/schemas/Uuid'
        ruleKey:
          type: string
          minLength: 1
        ruleVersion:
          type: integer
          minimum: 1
        careCategory:
          type: string
          minLength: 1
          maxLength: 100
        safetyTier:
          $ref: '#/components/schemas/RecommendationSafetyTier'
        state:
          $ref: '#/components/schemas/RecommendationState'
        urgency:
          $ref: '#/components/schemas/TaskUrgency'
        targetKind:
          $ref: '#/components/schemas/TaskTargetKind'
        targetGardenAreaMapObjectId:
          oneOf:
            - $ref: '#/components/schemas/Uuid'
            - type: 'null'
        targetPlantId:
          oneOf:
            - $ref: '#/components/schemas/Uuid'
            - type: 'null'
        windowStart:
          oneOf:
            - $ref: '#/components/schemas/Timestamp'
            - type: 'null'
        windowEnd:
          oneOf:
            - $ref: '#/components/schemas/Timestamp'
            - type: 'null'
        explanation:
          type: string
          minLength: 1
        supersedesCandidateId:
          oneOf:
            - $ref: '#/components/schemas/Uuid'
            - type: 'null'
          description: The prior candidate this one replaced or re-surfaced, when any — the walkable supersession/re-surfacing chain.
        presentedAt:
          oneOf:
            - $ref: '#/components/schemas/Timestamp'
            - type: 'null'
        revision:
          $ref: '#/components/schemas/Revision'
        createdAt:
          $ref: '#/components/schemas/Timestamp'
        updatedAt:
          $ref: '#/components/schemas/Timestamp'
    TodayRecommendation:
      type: object
      additionalProperties: false
      required:
        [
          id,
          gardenId,
          ruleKey,
          ruleVersion,
          careCategory,
          safetyTier,
          state,
          urgency,
          targetKind,
          targetGardenAreaMapObjectId,
          targetPlantId,
          windowStart,
          windowEnd,
          explanation,
          supersedesCandidateId,
          presentedAt,
          revision,
          createdAt,
          updatedAt,
          actionTitle,
          priorityScore,
          priorityFactors,
          evidence,
          targetDisplayName,
        ]
      description: |
        One Today item: every `Recommendation` field (declared flat, not
        via `allOf`, because both schemas forbid additional properties)
        plus everything the view alone renders — FR-24's proposed action,
        the re-derived priority score with its stored factors, the
        structured evidence, and the target's current display name.
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        gardenId:
          $ref: '#/components/schemas/Uuid'
        ruleKey:
          type: string
          minLength: 1
        ruleVersion:
          type: integer
          minimum: 1
        careCategory:
          type: string
          minLength: 1
          maxLength: 100
        safetyTier:
          $ref: '#/components/schemas/RecommendationSafetyTier'
        state:
          $ref: '#/components/schemas/RecommendationState'
        urgency:
          $ref: '#/components/schemas/TaskUrgency'
        targetKind:
          $ref: '#/components/schemas/TaskTargetKind'
        targetGardenAreaMapObjectId:
          oneOf:
            - $ref: '#/components/schemas/Uuid'
            - type: 'null'
        targetPlantId:
          oneOf:
            - $ref: '#/components/schemas/Uuid'
            - type: 'null'
        windowStart:
          oneOf:
            - $ref: '#/components/schemas/Timestamp'
            - type: 'null'
        windowEnd:
          oneOf:
            - $ref: '#/components/schemas/Timestamp'
            - type: 'null'
        explanation:
          type: string
          minLength: 1
        supersedesCandidateId:
          oneOf:
            - $ref: '#/components/schemas/Uuid'
            - type: 'null'
        presentedAt:
          oneOf:
            - $ref: '#/components/schemas/Timestamp'
            - type: 'null'
        revision:
          $ref: '#/components/schemas/Revision'
        createdAt:
          $ref: '#/components/schemas/Timestamp'
        updatedAt:
          $ref: '#/components/schemas/Timestamp'
        actionTitle:
          type: string
          minLength: 1
          description: The rule version's suggested action — also the title a conversion gives the task.
        priorityScore:
          type: integer
          minimum: 0
          maximum: 100
        priorityFactors:
          type: array
          items:
            $ref: '#/components/schemas/RecommendationPriorityFactor'
        evidence:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/RecommendationEvidence'
        targetDisplayName:
          type: [string, 'null']
          description: |
            The plant's current display name for plant targets; `null`
            for garden targets (the client knows its garden) and for
            garden-area targets, which no launch rule produces.
    TodayResult:
      type: object
      additionalProperties: false
      required: [gardenId, generatedAt, items]
      properties:
        gardenId:
          $ref: '#/components/schemas/Uuid'
        generatedAt:
          $ref: '#/components/schemas/Timestamp'
        items:
          type: array
          items:
            $ref: '#/components/schemas/TodayRecommendation'
    PostponeRecommendationRequest:
      type: object
      additionalProperties: false
      description: |
        `postponedUntil` is the user's optional re-surfacing horizon. No
        default is applied when absent or `null` — the engine falls back
        to the rule's own recurrence interval.
      properties:
        postponedUntil:
          oneOf:
            - $ref: '#/components/schemas/Timestamp'
            - type: 'null'
    ConvertRecommendationToTaskResult:
      type: object
      additionalProperties: false
      required: [recommendation, task]
      properties:
        recommendation:
          $ref: '#/components/schemas/Recommendation'
        task:
          $ref: '#/components/schemas/Task'

    # ------------------------------------------------------------------
    # Media (P6-API-01).
    #
    # Registration, authorized resumable upload sessions, client-triggered
    # completion verification, status, and short-lived authorized access.
    # `bucketName`/`objectKey` are deliberately never exposed on `Media`:
    # section 2, "Cloud Storage object names are opaque infrastructure
    # identifiers," and section 5, "Signed URLs and resumable session URLs
    # are not persisted as identity" — a client's only ways to reach the
    # underlying object are `uploadUrl` (from `MediaUploadSession`, at
    # registration) and the signed `url` (from `MediaAccess`, at download
    # time), never a bucket/key pair it could construct its own URL from.
    #
    # Source: architecture/media-storage-and-processing.md.
    # ------------------------------------------------------------------
...
  responses:
    BadRequest:
      description: The request is malformed or fails validation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication is missing or invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Authenticated but not authorized for this resource.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: |
        Not found, or existence intentionally concealed because the actor is not
        authorized to learn that the resource exists.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: A domain or synchronization conflict prevented the command.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    PreconditionFailed:
      description: The supplied revision precondition did not match the current revision.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
security:
  - firebaseIdToken: []
  - sessionCookie: []